# docs/swagger.yaml
basePath: /
definitions:
  internal_api.InvalidParam:
    properties:
      name:
        type: string
      reason:
        type: string
    type: object
  internal_api.LoginRequest:
    properties:
      api_key:
        type: string
    type: object
  internal_api.LoginResponse:
    properties:
      role:
        type: string
    type: object
  internal_api.MeResponse:
    properties:
      role:
        type: string
    type: object
  internal_api.ProblemDetail:
    properties:
      detail:
        description: Detail is a human-readable explanation specific to this occurrence.
        type: string
      instance:
        description: Instance is a URI reference that identifies the specific occurrence.
        type: string
      invalid_params:
        description: InvalidParams lists parameters that failed validation (RFC 9457
          extension).
        items:
          $ref: '#/definitions/internal_api.InvalidParam'
        type: array
      status:
        description: Status is the HTTP status code generated by the origin server.
        type: integer
      title:
        description: Title is a short, human-readable summary of the problem type.
        type: string
      trace_id:
        description: TraceID is the distributed trace identifier for this request.
        type: string
      type:
        description: Type is a URI reference that identifies the problem type.
        type: string
    type: object
  internal_api.swaggAccount:
    properties:
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
  internal_api.swaggTransaction:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
    type: object
  ledger_pro_internal_db.AccountType:
    enum:
    - asset
    - liability
    - equity
    - revenue
    - expense
    type: string
    x-enum-varnames:
    - AccountTypeAsset
    - AccountTypeLiability
    - AccountTypeEquity
    - AccountTypeRevenue
    - AccountTypeExpense
  ledger_pro_internal_db.PostingDirection:
    enum:
    - debit
    - credit
    type: string
    x-enum-varnames:
    - PostingDirectionDebit
    - PostingDirectionCredit
  ledger_pro_internal_service.BalanceResponse:
    properties:
      account_id:
        type: integer
      currency:
        type: string
      net_balance:
        type: integer
      total_credits:
        type: integer
      total_debits:
        type: integer
    type: object
  ledger_pro_internal_service.CreateAccountRequest:
    properties:
      account_name:
        type: string
      account_type:
        $ref: '#/definitions/ledger_pro_internal_db.AccountType'
      currency:
        type: string
      metadata:
        items:
          type: integer
        type: array
    type: object
  ledger_pro_internal_service.CreateTransactionRequest:
    properties:
      description:
        type: string
      idempotency_key:
        type: string
      metadata:
        items:
          type: integer
        type: array
      postings:
        items:
          $ref: '#/definitions/ledger_pro_internal_service.PostingEntry'
        type: array
    type: object
  ledger_pro_internal_service.PostingEntry:
    properties:
      account_id:
        type: integer
      amount:
        type: integer
      direction:
        $ref: '#/definitions/ledger_pro_internal_db.PostingDirection'
    type: object
host: ledger-pro-api.onrender.com
info:
  contact: {}
  description: A FAANG-grade double-entry ledger system.
  title: LedgerPro API
  version: "1.0"
paths:
  /v1/accounts:
    get:
      description: Retrieves a paginated list of accounts
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - default: 50
        description: Number of items to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api.swaggAccount'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
      security:
      - ApiKeyAuth: []
      summary: List Accounts
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Creates a new double-entry ledger account
      parameters:
      - description: Account details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ledger_pro_internal_service.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.swaggAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
      security:
      - ApiKeyAuth: []
      summary: Create Account
      tags:
      - Accounts
  /v1/accounts/{id}:
    get:
      description: Retrieves an account by its ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.swaggAccount'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
      security:
      - ApiKeyAuth: []
      summary: Get Account
      tags:
      - Accounts
  /v1/accounts/{id}/balance:
    get:
      description: Computes the real-time balance of an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ledger_pro_internal_service.BalanceResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
      security:
      - ApiKeyAuth: []
      summary: Get Account Balance
      tags:
      - Accounts
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and sets a secure session cookie
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
      summary: Login
      tags:
      - Authentication
  /v1/auth/logout:
    post:
      description: Revokes the active session and clears the session cookie
      responses:
        "204":
          description: No Content
      summary: Logout
      tags:
      - Authentication
  /v1/auth/me:
    get:
      description: Returns the role of the currently authenticated session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
      security:
      - ApiKeyAuth: []
      summary: Get Current User
      tags:
      - Authentication
  /v1/health:
    get:
      description: Checks connectivity to Postgres, Redis, RabbitMQ, and R2.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
      summary: Health Check
      tags:
      - Observability
  /v1/transactions:
    get:
      description: Retrieves a paginated list of transactions
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - default: 50
        description: Number of items to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api.swaggTransaction'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
      security:
      - ApiKeyAuth: []
      summary: List Transactions
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: Processes a double-entry transaction
      parameters:
      - description: Idempotency Key (UUID)
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Transaction details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ledger_pro_internal_service.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.swaggTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
      security:
      - ApiKeyAuth: []
      summary: Create Transaction
      tags:
      - Transactions
  /v1/transactions/{id}:
    get:
      description: Retrieves a transaction by its ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.swaggTransaction'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ProblemDetail'
      security:
      - ApiKeyAuth: []
      summary: Get Transaction
      tags:
      - Transactions
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"